+++ app/Http/Controllers/StudentsController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Student;
 use App\Models\School;
 use App\Models\Classes;
+use App\Models\Role;
 use App\Models\EmployeeSubject;
 use App\Models\StudentSubject;
 use Auth;
@@ -13,8 +14,10 @@ use Auth;
 class StudentsController extends Controller
 {
     public function index(){
+        $this->authorize('create',  Student::class);
+
         $schools = School::all()->where('deleted_at', NULL);
-        if(Auth::user()->role_id != 4){
+        if(Auth::user()->role_id != Role::IS_CLASSTEACHER){
             $classes = Classes::all()->where('deleted_at', NULL)->where('school_id', Auth::user()->school_id);
         }else{
             //allow class teacher to add students belonging to their class only
@@ -23,6 +26,8 @@ class StudentsController extends Controller
         return view('students/addStudent', ['schools'=>$schools, 'classes'=>$classes]);
     }
     public function store(Request $request){
+        $this->authorize('create',  Student::class);
+
         $request->validate([
             'fname' => 'required',
             'lname' => 'required',
@@ -43,17 +48,19 @@ class StudentsController extends Controller
         ]);
 
         return redirect('/viewstudents')->with('message', 'Student added successfully!');
-
     }
+    
     public function viewStudents(){
-        //display all students only when admin is logged in
-        if(Auth::user()->role_id == 1){
+        $this->authorize('viewAny',  Student::class);
+
+        //display all students only when admin is logged in        
+        if(Auth::user()->role_id == Role::IS_SUPERADMIN){
             $students = Student::where('deleted_at', NULL)->get();
            
             return view('students/viewStudents', ['students'=> $students]);
         }
         //display students in the class where the logged in teacher is a classteacher only
-        elseif(Auth::user()->role_id == 4){
+        elseif(Auth::user()->role_id == Role::IS_CLASSTEACHER){
             $students = Student::select("*")
                     ->whereIn('class_id', Classes::select('id')->where('deleted_at', NULL)->where('class_teacher', Auth::user()->id)->get())
                     ->get();
@@ -78,19 +85,23 @@ class StudentsController extends Controller
         }
     }
     public function viewStudentsTaughtByEmployee($id){
+        $this->authorize('viewAny',  Student::class);
+
         //select all students who are in the class and do the subject taught by the teacher
         $students = Student::select('*')->where('class_id', EmployeeSubject::find($id)->class_id)->whereIn('id', StudentSubject::select('student_id')->where('subject_id', EmployeeSubject::find($id)->subject_id)->get())->get();
         return view('students/viewStudentsToAddMarks', ['students'=>$students, 'subject'=>EmployeeSubject::find($id)->subject_id]);
     }
     public function viewStudentsToAddMarks($id){
+        $this->authorize('viewAny',  Student::class);
+
         //display all students only when admin is logged in
-        if(Auth::user()->role_id == 1){
+        if(Auth::user()->role_id == Role::IS_SUPERADMIN){
             $students = Student::where('deleted_at', NULL)->get();
            
             return view('students/viewStudents', ['students'=> $students]);
         }
         //display students in the class where the logged in teacher is a classteacher only
-        elseif(Auth::user()->role_id == 4){
+        elseif(Auth::user()->role_id == Role::IS_CLASSTEACHER){
             $students = Student::select("*")
                     ->whereIn('class_id', Classes::select('id')->where('deleted_at', NULL)->where('class_teacher', Auth::user()->id)->get())
                     ->get();
@@ -121,14 +132,18 @@ class StudentsController extends Controller
         return view('students/viewStudentstoAddMarks', ['students'=>$students, 'subject'=>EmployeeSubject::find($id)->subject_id]);
     }
     
-    public function edit($id){
+    public function edit($id, Student $editStudent){
+        $this->authorize('update',  $editStudent);
+
         $student = Student::find($id);
-        $classes = Classes::all()->where('deleted_at', NULL);
+        $classes = Classes::all()->where('deleted_at', NULL)->where('school_id', Auth::user()->school_id);
 
         return view('students/editstudent', ['student'=>$student, 'classes' => $classes]);
     }
 
-    public function update(Request $request, $id){
+    public function update(Request $request, $id, Student $editStudent){
+        $this->authorize('update',  $editStudent);
+
         $request->validate([
             'fname' => 'required',
             'lname' => 'required',
@@ -154,8 +169,10 @@ class StudentsController extends Controller
         return redirect('/viewstudents')->with('message', 'Student updated successfully!');
     }
 
-    public function destroy($id)
+    public function destroy($id, Student $deleteStudent)
     {
+        $this->authorize('delete',  $deleteStudent);
+
         $student = Student::find($id)->delete();
         return redirect('/viewstudents')->with('message', 'Student deleted successfully!');
     }
@@ -172,18 +189,24 @@ class StudentsController extends Controller
 
     //softDeletes students
     public function trashedStudents(){
+        $this->authorize('restore',  Student::class);
+
         $students = Student::onlyTrashed()->get();
         return view('students/trashedStudents', compact('students'));
     }
 
     //restore deleted students
     public function restoreStudent($id){
+        $this->authorize('restore',  Student::class);
+
         Student::whereId($id)->restore();
         return back();
     }
 
     //restore all deleted students
     public function restoreStudents(){
+        $this->authorize('restore',  Student::class);
+
         Student::onlyTrashed()->restore();
         return back();
     }
