+++ app/Http/Controllers/SubjectsController.php
@@ -9,11 +9,15 @@ use App\Models\Subject;
 class SubjectsController extends Controller
 {
     public function index(){
+        $this->authorize('create',  Subject::class);
+
         $categories = SubjectCategories::all()->where('deleted_at', NULL);
 
         return view('subjects/addSubject', ['categories'=> $categories ]);
     }
     public function store(Request $request){
+        $this->authorize('create',  Subject::class);
+
         $request->validate([
             'name' => 'required',
             'category'=>'required'
@@ -27,19 +31,25 @@ class SubjectsController extends Controller
         return redirect('/viewsubjects')->with('message', 'Subject added successfully!');
     }
     public function viewSubjects(){
+        $this->authorize('viewAny',  Subject::class);
+
         $subjects = Subject::where('deleted_at', NULL)->get();
 
         return view('subjects/viewSubjects', ['subjects'=> $subjects]);
     }
 
-    public function edit($id){
+    public function edit($id, Subject $editSubject){
+        $this->authorize('update',  $editSubject);
+
         $subject = Subject::find($id);
         $categories = SubjectCategories::all()->where('deleted_at', NULL);
 
         return view('subjects/editSubject', ['subject'=> $subject,'categories'=> $categories]);
     }
 
-    public function update(Request $request, $id){
+    public function update(Request $request, $id, Subject $editSubject){
+        $this->authorize('update',  $editSubject);
+
         $request->validate([
             'name' => 'required',
             'category' => 'required',
@@ -55,8 +65,9 @@ class SubjectsController extends Controller
         return redirect('/viewsubjects')->with('message', 'Subject updated successfully!');
     }
 
-    public function destroy($id)
+    public function destroy($id, Subject $deleteSubject)
     {
+        $this->authorize('delete',  $deleteSubject);
         $subject = Subject::find($id)->delete();
         
         return redirect('/viewsubjects')->with('message', 'Subject deleted successfully!');
@@ -73,18 +84,24 @@ class SubjectsController extends Controller
 
     //softDeletes subjects
     public function trashedSubjects(){
+        $this->authorize('restore', Subject::class);
+
         $subjects = Subject::onlyTrashed()->get();
         return view('subjects/trashedSubjects', compact('subjects'));
     }
 
     //restore deleted subjects
     public function restoreSubject($id){
+        $this->authorize('restore', Subject::class);
+
         Subject::whereId($id)->restore();
         return back();
     }
 
     //restore all deleted subjects
     public function restoreSubjects(){
+        $this->authorize('restore', Subject::class);
+        
         Subject::onlyTrashed()->restore();
         return back();
     }
